+++ src/components/layout/Layout.tsx
@@ -1,6 +1,6 @@
 import { FC, ReactNode } from 'react'
 import { Footer } from './Footer'
-import { Header } from './header/Header'
+import { Header } from './Header'
 
 type Props = {
   children: ReactNode
