+++ src/components/post/PostFolderList.tsx
@@ -1,8 +1,8 @@
 import { FC } from 'react'
 
 import { AiFillFolder as AiFillFolderIcon } from 'react-icons/ai'
+import { CreateFolderButton } from 'components/shares/CreateFolderButton'
 
-import { CreateFolderButton } from 'components/shares/button'
 import { useGetApi } from 'hooks/useApi'
 import { Folder } from 'types/bookmark'
 import { Post } from 'types/post'
