+++ src/components/post/PostMenuButton.tsx
@@ -2,8 +2,8 @@ import { FC, useCallback, useMemo, useState } from 'react'
 
 import { BsThreeDots as BsThreeDotsIcon } from 'react-icons/bs'
 
-import { FolderList } from './FolderList'
-import { DropDownMenu } from 'components/shares/DropDownMenu'
+import { FolderList } from './PostFolderList'
+import { DropDownMenu } from 'components/shares/base/DropDownMenu'
 import { useGetApi } from 'hooks/useApi'
 import { Post } from 'types/post'
 import { User } from 'types/user/user'
