--- src/components/shares/button/index.ts
@@ -1,3 +1,0 @@
-export { Button } from './Button'
-export { JumpToCreatePostButton } from './JumpToCreatePostButton'
-export { CreateFolderButton } from './CreateFolderButton'
