+++ src/pages/bookmark.tsx
@@ -3,8 +3,8 @@ import { useSearchParams } from 'next/navigation'
 import { useMemo } from 'react'
 import { BookmarkFolderList } from 'components/bookmark/BookmarkFolderList'
 import { Layout } from 'components/layout/Layout'
-import { PostItem } from 'components/post/Item/PostItem'
-import { CreateFolderButton } from 'components/shares/button'
+import { PostItem } from 'components/post/PostItem'
+import { CreateFolderButton } from 'components/shares/CreateFolderButton'
 import { useRequireLogin } from 'hooks/login/useRequireLogin'
 import { useGetApi } from 'hooks/useApi'
 import { Folder, BookmarkPosts } from 'types/bookmark'
