+++ src/pages/users/[id].tsx
@@ -3,9 +3,9 @@ import { useSearchParams } from 'next/navigation'
 import { useState } from 'react'
 
 import { Layout } from 'components/layout/Layout'
-import { PostItem } from 'components/post/Item/PostItem'
-import { Avatar } from 'components/shares/Avatar'
-import { Button } from 'components/shares/button'
+import { PostItem } from 'components/post/PostItem'
+import { Avatar } from 'components/shares/base/Avatar'
+import { Button } from 'components/shares/base/Button'
 import { useGetApi } from 'hooks/useApi'
 import { UserPosts } from 'types/post'
 import { User } from 'types/user/user'
